+++ src/rules/unit-case/__tests__/index.js
@@ -83,6 +83,12 @@ testRule(rule, {
   }, {
     code: "a { margin: 13xpx; }",
     description: "work with unknown units",
+  }, {
+    code: "@media (min-width: 10px) {}",
+    description: "check @media",
+  }, {
+    code: "@media (min-width: 10px)\n  and (max-width: 20px) {}",
+    description: "check complex @media",
   } ],
 
   reject: [ {
@@ -136,6 +142,24 @@ testRule(rule, {
     description: "work with unknown units",
     line: 1,
     column: 13,
+  }, {
+    code: "@media (min-width: 13PX) {}",
+    message: messages.expected("PX", "px"),
+    description: "@media",
+    line: 1,
+    column: 20,
+  }, {
+    code: "@media (min-width: 10px)\n  and (max-width: 20PX) {}",
+    message: messages.expected("PX", "px"),
+    description: "complex @media",
+    line: 2,
+    column: 19,
+  }, {
+    code: "@media (width < 10.01REM) {}",
+    message: messages.expected("REM", "rem"),
+    description: "media feature range",
+    line: 1,
+    column: 17,
   } ],
 })