+++ src/rules/unit-case/index.js
@@ -1,4 +1,5 @@
 import {
+  atRuleParamIndex,
   declarationValueIndex,
   getUnitFromValueNode,
   report,
@@ -24,12 +25,12 @@ export default function (expectation) {
     })
     if (!validOptions) { return }
 
-    root.walkDecls(decl => {
-      valueParser(decl.value).walk((node) => {
+    function check(node, value, getIndex) {
+      valueParser(value).walk((valueNode) => {
         // Ignore wrong units within `url` function
-        if (node.type === "function" && node.value.toLowerCase() === "url") { return false }
+        if (valueNode.type === "function" && valueNode.value.toLowerCase() === "url") { return false }
 
-        const unit = getUnitFromValueNode(node)
+        const unit = getUnitFromValueNode(valueNode)
 
         if (!unit) { return }
 
@@ -38,13 +39,20 @@ export default function (expectation) {
         if (unit === expectedUnit) { return }
 
         report({
+          index: getIndex(node) + valueNode.sourceIndex,
           message: messages.expected(unit, expectedUnit),
-          node: decl,
-          index: declarationValueIndex(decl) + node.sourceIndex,
+          node,
           result,
           ruleName,
         })
       })
-    })
+    }
+
+    root.walkAtRules(/^media$/i, atRule =>
+      check(atRule, atRule.params, atRuleParamIndex)
+    )
+    root.walkDecls(decl =>
+      check(decl, decl.value, declarationValueIndex)
+    )
   }
 }